software/src/requirements.txt
pandas==2.2.3
polars-lts-cpu==1.33.1
polars-ds-lts-cpu==0.10.2

